feat - 유저 포인트 받기 api 구현

+++ src/main/java/sopt/org/hmh/global/auth/redis/RefreshToken.java
@@ -21,5 +21,5 @@ public class RefreshToken {
     private String token;
 
     @TimeToLive
-    private Integer expiration;
+    private Long expiration;
 }
